Suggest/Accuse Buttons
Added commands for suggest/accusation buttons.  Had to bind an observable object collection to the combo boxes instead of hard coding them to get it to work.  Also broke the pop-up functionality for those buttons, but we can easily fix it or delete those functions since the controller should be handling those calls.